Call HandleTransactionError when transaction removed from pending pool without being committed

+++ test/acceptance/main_test.go
@@ -7,9 +7,11 @@ import (
 )
 
 func TestMain(m *testing.M) {
-	logs := config.GetProjectSourceRootPath() + "/logs/acceptance/"
-	os.RemoveAll(logs)
-	os.MkdirAll(logs, 0755)
+	if os.Getenv("NO_LOG_STDOUT") == "true" {
+		logs := config.GetProjectSourceRootPath() + "/logs/acceptance/"
+		os.RemoveAll(logs)
+		os.MkdirAll(logs, 0755)
+	}
 
 	os.Exit(m.Run())
 }
